MSBlog/MSBlog/blog/serializers.py
from .models import Post, Tag

from rest_framework import serializers


class PostSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Post
        fields = ('title', 'author', 'reading')
